wip(module:collapse): refactor collapse
