+++ components/collapse/demo/basic.ts
@@ -3,11 +3,11 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'nz-demo-collapse-basic',
   template: `
-    <nz-collapseset>
-      <nz-collapse *ngFor="let panel of panels" [nzTitle]="panel.name" [nzActive]="panel.active" [nzDisabled]="panel.disabled">
+    <nz-collapse>
+      <nz-collapse-panel *ngFor="let panel of panels" [nzHeader]="panel.name" [nzActive]="panel.active" [nzDisabled]="panel.disabled">
         <p style="margin:0;">A dog is a type of domesticated animal. Known for its loyalty and faithfulness, it can be found as a welcome guest in many households across the world.</p>
-      </nz-collapse>
-    </nz-collapseset>
+      </nz-collapse-panel>
+    </nz-collapse>
   `,
   styles  : []
 })
