+++ components/collapse/demo/noarrow.ts
@@ -3,11 +3,11 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'nz-demo-collapse-noarrow',
   template: `
-    <nz-collapseset>
-      <nz-collapse *ngFor="let panel of panels" [nzTitle]="panel.name" [nzActive]="panel.active" [nzDisabled]="panel.disabled" [nzShowArrow]="panel.arrow">
+    <nz-collapse>
+      <nz-collapse-panel *ngFor="let panel of panels" [nzHeader]="panel.name" [nzActive]="panel.active" [nzDisabled]="panel.disabled" [nzShowArrow]="panel.arrow">
         <p style="margin:0;">A dog is a type of domesticated animal. Known for its loyalty and faithfulness, it can be found as a welcome guest in many households across the world.</p>
-      </nz-collapse>
-    </nz-collapseset>
+      </nz-collapse-panel>
+    </nz-collapse>
   `,
   styles  : []
 })
