+++ components/collapse/doc/index.zh-CN.md
@@ -15,18 +15,19 @@ cols: 1
 
 ## API
 
-### Collapse
+### nz-collapse
 
 | 参数 | 说明 | 类型 | 默认值 |
 | --- | --- | --- | --- |
-| activeKey | 当前激活 tab 面板的 key | string\[]｜string | 默认无，accordion模式下默认第一个元素 |
-| defaultActiveKey | 初始化选中面板的 key | string | 无 |
-| onChange | 切换面板的回调 | Function | 无 |
+| nzAccordion | 是否每次只打开一个tab | boolean | false |
+| nzBordered | 是否有边框 | boolean | true |
 
-### Collapse.Panel
+### nz-collapse-panel
 
 | 参数 | 说明 | 类型 | 默认值 |
 | --- | --- | --- | --- |
-| disabled | 禁用后的面板展开与否将无法通过用户交互改变 | boolean | false |
-| header | 面板头内容 | string｜ReactNode | 无 |
-| key | 对应 activeKey | string | 无 |
+| nzDisabled | 禁用后的面板展开与否将无法通过用户交互改变 | boolean | false |
+| nzHeader | 面板头内容 | string｜ `TemplateRef<void>` | 无 |
+| nzShowArrow | 是否展示箭头 | boolean | true |
+| nzActive | 面板是否展开，可双向绑定 | boolean | 无 |
+| nzActiveChange | 面板展开回调 | (nzActive:boolean)=>{} | 无 |
