+++ components/collapse/public-api.ts
@@ -1,0 +1,3 @@
+export * from './nz-collapse-panel.component';
+export * from './nz-collapse.component';
+export * from './nz-collapse.module';
