+++ components/index.ts
@@ -8,9 +8,9 @@ import { NzBadgeModule } from './badge';
 import { NzBreadCrumbModule } from './breadcrumb/nz-breadcrumb.module';
 import { NzButtonModule } from './button';
 import { NzCardModule } from './card';
-import { NzCarouselModule } from './carousel/nz-carousel.module';
+import { NzCarouselModule } from './carousel';
 import { NzCheckboxModule } from './checkbox';
-import { NzCollapseModule } from './collapse/nz-collapse.module';
+import { NzCollapseModule } from './collapse';
 import { NzDividerModule } from './divider/nz-divider.module';
 import { NzDropdownService, NzDropDownModule } from './dropdown';
 import { NzFormModule } from './form/nz-form.module';
